servicio_mototaxis/requirements.txt
fastapi>=0.100.0
uvicorn==0.29.0
websockets==11.0.3
python-dotenv>=1.0.0
pydantic>=2.0
pydantic-settings>=2.0.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.7.0
redis>=4.3.0
pika>=1.3.0
httptools
watchfiles
pydantic[email]
python-multipart